reset

+++ recolte.py
@@ -28,18 +28,19 @@ l = liste_stations_ville("lyon")
 
 r = []
 for k in l :
-    tmp = {}
-    tmp["number"] = k["number"]
-    tmp["totalStands"] = k["totalStands"]
+    tmp = (k["number"], k["totalStands"]["availabilities"]["bikes"])
     r.append(tmp)
+
+r = sorted(r, key=lambda x: x[0])
+r = [i[1] for i in r]
 
 import time
 nom = "data/"+str(int(time.time()))+".json"
 with open(nom, "w") as f :
     json.dump(r, f)
-
-import os
-os.system("git add " + nom)
-os.system("git commit -m 'la routine' " + nom)
-os.system("git push")
-os.system("rm " + nom)
+if True :
+    import os
+    os.system("git add " + nom)
+    os.system("git commit -m 'la routine' " + nom)
+    os.system("git push")
+    os.system("rm " + nom)
